Relocate workout-form types and add tsconfig for tests
